+++ src/components/index.js
@@ -5,7 +5,6 @@ import ButtonGroup from 'components/ButtonGroup';
 import Checkbox from 'components/Checkbox';
 import Chip from 'components/Chip';
 import DateInput from 'components/DateInput';
-import DecoratedInput from 'components/DecoratedInput';
 import Dialog from 'components/Dialog';
 import DropdownButton from 'components/DropdownButton';
 import Flex from 'components/Flex';
@@ -13,7 +12,6 @@ import FlexItem from 'components/FlexItem';
 import FloatingAction from 'components/FloatingAction';
 import Hscroll from 'components/Hscroll';
 import IconButton from 'components/IconButton';
-import IconInput from 'components/IconInput';
 import InputGroup from 'components/InputGroup';
 import Menu from 'components/Menu';
 import MenuItem from 'components/MenuItem';
@@ -35,7 +33,6 @@ export {
   Checkbox,
   Chip,
   DateInput,
-  DecoratedInput,
   Dialog,
   DropdownButton,
   Flex,
@@ -43,7 +40,6 @@ export {
   FloatingAction,
   Hscroll,
   IconButton,
-  IconInput,
   InputGroup,
   Menu,
   MenuItem,